연결 부분 모두 수정
