+++ src/com/goods/db/listDAO.java
@@ -19,20 +19,42 @@ public class listDAO extends DBconnection{
 	
 	private ResultSet rs = null;
 	private String sql = "";
-	
+	private Connection conn = null;
+	private PreparedStatement pstmt = null;
+
+	varlist vars = new varlist();
+	private String databasename = vars.getDatabasename();
 	varlist var = new varlist();
 	String tablename = var.getGoodslistTablename(); 
 	
 	DBconnection con = new DBconnection();
 	setGoodsTool setTool = new setGoodsTool();
+	public void closeDB() {
+		try {
+			if (rs != null) {
+				rs.close();
+			}
+			if (pstmt != null) {
+				pstmt.close();
+			}
+			if (conn != null) {
+				conn.close();
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+	}
 	public List getGoodsList() {
 		List goodsList = new ArrayList();
 		try {
 			
 			
 			sql = "select * from "+tablename;
-			rs = con.selsql(sql);
-
+			conn = getConnection();
+			
+			pstmt = conn.prepareStatement(sql);
+			rs = pstmt.executeQuery();
+			
 			while (rs.next()) {
 				
 				
@@ -51,7 +73,7 @@ public class listDAO extends DBconnection{
 		} catch (SQLException e) {
 			e.printStackTrace();
 		} finally {
-			con.closeDB();
+			
 			closeDB();
 		}
 
@@ -77,7 +99,10 @@ public class listDAO extends DBconnection{
 									+ " cosCategory like '%"+item+"%' ";
 			System.out.println("sql = "+  sql);
 			
-			rs = con.selsql(sql);
+			conn = getConnection();
+			
+			pstmt = conn.prepareStatement(sql);
+			rs = pstmt.executeQuery();
 			
 			while (rs.next()) {
 				GoodsDTO goods = new GoodsDTO();
@@ -132,7 +157,10 @@ public class listDAO extends DBconnection{
 					}
 			}
 			System.out.println(sql);
-			rs = con.selsql(sql);
+			conn = getConnection();
+			
+			pstmt = conn.prepareStatement(sql);
+			rs = pstmt.executeQuery();
 			
 			while (rs.next()) {
 				GoodsDTO goods = new GoodsDTO();
@@ -184,8 +212,10 @@ public class listDAO extends DBconnection{
 						sql = "select * from "+tablename + "order by cosNum desc limit " + startRow +" , "+ pageSize;
 					}
 			}
-			System.out.println(sql);
-			rs = con.selsql(sql);
+			conn = getConnection();
+			
+			pstmt = conn.prepareStatement(sql);
+			rs = pstmt.executeQuery();
 			
 			while (rs.next()) {
 				GoodsDTO goods = new GoodsDTO();
@@ -225,7 +255,10 @@ public class listDAO extends DBconnection{
 			
 			
 				
-				rs = con.selsql(sql);
+				conn = getConnection();
+				
+				pstmt = conn.prepareStatement(sql);
+				rs = pstmt.executeQuery();
 				
 				while (rs.next()) {
 					GoodsDTO goods = new GoodsDTO();
@@ -258,8 +291,10 @@ public class listDAO extends DBconnection{
 			try {
 				
 				sql ="select distinct cosBrand from cos_list";
-				rs = con.selsql(sql);
+				conn = getConnection();
 				
+				pstmt = conn.prepareStatement(sql);
+				rs = pstmt.executeQuery();
 				while (rs.next()) {
 					testList.add(rs.getString("cosBrand"));
 				} 
@@ -283,8 +318,10 @@ public class listDAO extends DBconnection{
 			try {
 				
 				sql ="select distinct cosCategory from cos_list";
-				rs = con.selsql(sql);
+				conn = getConnection();
 				
+				pstmt = conn.prepareStatement(sql);
+				rs = pstmt.executeQuery();
 				while (rs.next()) {
 					testList.add(rs.getString("cosCategory"));
 				} 
@@ -308,7 +345,10 @@ public class listDAO extends DBconnection{
 			try {
 				
 				sql ="select distinct cosSkinType from cos_list";
-				rs = con.selsql(sql);
+				conn = getConnection();
+				
+				pstmt = conn.prepareStatement(sql);
+				rs = pstmt.executeQuery();
 				
 				while (rs.next()) {
 					testList.add(rs.getString("cosSkinType"));
