+++ src/com/notice/db/noticeUpdate.java
@@ -1,17 +1,29 @@
 package com.notice.db;
 
+import java.sql.Connection;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 
+import javax.naming.Context;
+import javax.naming.InitialContext;
+import javax.naming.NamingException;
+import javax.sql.DataSource;
 
 import com.var.list.varlist;
 
 
-public class noticeUpdate extends DBconnection{
+public class noticeUpdate {
 	varlist var = new varlist();
 	private String tablename = var.getnoticelistTablename();
+	private PreparedStatement pstmt = null;
+	private ResultSet rs = null;
+	private String sql = "";
+	private Connection conn = null;
+	private String databasename = var.getDatabasename();
 	
 	public void doit(noticeDTO noti){
-		DBconnection con = new DBconnection();
+		
 		setnoticetool notit = new setnoticetool();
 		ResultSet rs = null;
 		String sql = "";
@@ -28,7 +40,10 @@ public class noticeUpdate extends DBconnection{
 				+ " noticeRealImgName = '"+noti.getNoticeRealImgName() +"'"
 				+ " where  noticeNum = "+noti.getNoticeNum()+" ";
 				System.out.println("수정 sql = "+ sql );
-				con.upsql(sql);
+				conn = getConnection();
+				
+				pstmt = conn.prepareStatement(sql);
+				pstmt.executeUpdate();
 				
 
 				System.out.println("DAO : 공지 등록 완료");
@@ -36,11 +51,47 @@ public class noticeUpdate extends DBconnection{
 			} catch (Exception e) {
 				e.printStackTrace();
 			} finally {
-				con.closeDB();
+				closeDB();
 				
 			}
 
 		
 		
+	}
+	public void closeDB() {
+		try {
+			if (rs != null) {
+				rs.close();
+			}
+			if (pstmt != null) {
+				pstmt.close();
+			}
+			if (conn != null) {
+				conn.close();
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+	}
+	protected Connection getConnection() {
+		try {
+			// Context 객체를 생성 (프로젝트 정보를 가지고있는객체)
+			Context initCTX = new InitialContext();
+			// DB연동 정보를 불러오기(context.xml)
+			
+			DataSource ds = (DataSource) initCTX.lookup("java:comp/env/jdbc/"+ databasename);
+
+			conn = ds.getConnection();
+
+			System.out.println(" 드라이버 로드, 디비연결 성공! ");
+			System.out.println(conn);
+
+		} catch (NamingException e) {
+			e.printStackTrace();
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		
+		return conn;
 	}
 }
